+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 v0.2.0 (in development)
 -----------------------
+- Added an `inversions()` method for returning the inversion number of a
+  permutation (based on a patch by @drocta)
 
 v0.1.1 (2018-09-18)
 -------------------